first implementation ligt price service call

+++ domain/model/cost.go
@@ -3,6 +3,6 @@ package model
 import "time"
 
 type Cost struct {
-	Value uint
+	Value float64
 	Date  time.Time
 }
